V1.1.9k64 pre
Changes:
1. Relational operators < > <= >= == <> now operate on components of list, similar to Plus and Times. Thus Equal/Unequal for two  lists with identical shape does not return True nor False. For that purpoes use === or <=>.

2. ExternalMap["TRACK"..] accepts CompiledMap for code compiled by Fortran. Descriptions will follow.

+++ src/tfsolvemember.f
@@ -44,7 +44,7 @@
         endif
         return
       elseif(list%head%k .eq. ktfoper+mtfslot)then
-        call tfslot(int8(mtfslot),list,kx,.false.,irtc)
+        call tfslot(mtfslot,list,kx,.false.,irtc)
         if(irtc .gt. 0 .and. ierrorprint .ne. 0)then
           call tfreseterror
           irtc=-1
